move note.tsx into components folder, add droppable area for notes
