+++ app/server/[serverId]/notes/_components/note.tsx
