+++ package.json
@@ -23,10 +23,10 @@
     "@hookform/resolvers": "^3.3.4",
     "@paralleldrive/cuid2": "^2.2.2",
     "@prisma/client": "^5.11.0",
-    "axios": "^1.6.8",
     "@tiptap/pm": "^2.3.0",
     "@tiptap/react": "^2.3.0",
     "@tiptap/starter-kit": "^2.3.0",
+    "axios": "^1.6.8",
     "bcryptjs": "^2.4.3",
     "clsx": "^2.1.0",
     "crypto-browserify": "^3.12.0",
